+++ src/data/index.js
@@ -1,0 +1,11 @@
+import { List } from "immutable";
+
+export const initialState = List.of(
+    {
+        slug: "microsoft-catalogo",
+        image: "http://www.hdwallpapers.org/wallpapers/caribbean-sea-beautiful-clouds-1920x1080.jpg",
+        name: "Microsoft Catálogo",
+        description: "Catálogo de Produtos desenvolvido para uma revenda Microsoft, seguindo a interface padrão do Window 8",
+        link: "http://microsoftcatalogo.azurewebsites.net"
+    }
+);