+++ src/reducers/App.js
@@ -1,4 +1,6 @@
-export default function(state = {}, action = {}) {
+import { initialState } from '../data/';
+
+export default function(state = initialState, action = {}) {
 
     //Default return
     return state;